+++ src/rapidfuzz/distance/__init__.py
@@ -11,4 +11,4 @@ Opcodes = _fallback_import(_mod, "Opcodes")
 ScoreAlignment = _fallback_import(_mod, "ScoreAlignment")
 MatchingBlock = _fallback_import(_mod, "MatchingBlock")
 
-from . import Hamming, Indel, Jaro, JaroWinkler, Levenshtein, LCSseq
+from . import Hamming, Indel, Jaro, JaroWinkler, Levenshtein, LCSseq, DamerauLevenshtein