+++ src/rapidfuzz/generate.sh
@@ -18,3 +18,4 @@ generate_cython distance/Hamming_cpp
 generate_cython distance/Indel_cpp
 generate_cython distance/LCSseq_cpp
 generate_cython distance/Levenshtein_cpp
+generate_cython distance/DamerauLevenshtein_cpp